third_party/blink/web_tests/fast/forms/select/menulist-popup-type-ahead-style-change.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../forms/resources/picker-common.js"></script>

<select>
  <option>A</option>
  <option>B</option>
  <option>C</option>
</select>

<script>
// Tests <select> style change when popup is visible after selection is changed
// by keyboard typeahead. The selection should not change when style changes.
window.onload = () => {
  const select = document.querySelector('select');
  openPickerAppearanceOnly(select, () => {
    eventSender.keyDown('C', null);
    select.offsetTop;
    select.style.backgroundColor = 'yellow';
    assert_equals(internals.selectMenuListText(select), 'C');
  });
};
</script>
